moved LineithVerticalBars out in new component
